Add files via upload
Documentation:

This software program is called the Integrated Tax Management tool, and its purpose is to facilitate tax calculation and management for several assets, particularly automobiles (busess) and constructions (houses). It offers organized classes and methods which are used to create buses and houses, get the details of assets and compute taxes with regards to different parameters of the assets.

The system enables the users to input information for different types of asset (buses & Houses ) and arrive at their respective tax values. However, taxation availed factors do vary across the different asset types;

Buses: Tax is calculated from the market value of the bus and seat numbers available.
Houses: Tax is considered based on geolocation (urban or rural), valuation of the house & square footage.
It is a console application utilized in a menu-based system and provides for the entry of assets with the display of tax per asset as well as a brief report on the taxes. The application will permanently stop working as soon as the last problem concerning taxes is solved.

Topics Covered are the following:

Classes and Objects: Designed base and derived classes for the buses and houses with attributes such as registration number, value and area among others.
Inheritance: Inheritance was employed to picture the classes in regard to vehicles and houses with the appropriate methods and properties.
Interfaces: The Taxable interface was applied for purposes of implementing the taxvalue() method in all classes with taxable objects, without appealing for its definition in asset classes.
Scanner for Input Handling: In order to take user inputs, a Scanner was used enabling the user to input asset details at run time.
Encapsulation: How some properties pertaining to assets were included in classes and access to them was through getter functions.
Polymorphism: The taxvalue() method was overridden in the classes TaxableBus and TaxableHouse showcasing different concepts where custom tax values were used.
Exception Handling: Program was aborted safely by handling exceptions.